fix ke 1

+++ database/migrations/2022_12_21_042825_create_barangs_table.php
@@ -17,7 +17,7 @@ class CreateBarangsTable extends Migration
             $table->id();
             $table->integer('kategori_id');
             $table->string('name');
-            $table->integer('berat');
+            $table->float('berat');
             $table->integer('harga_id');
             $table->timestamps();
         });
